refactor: comment api/hooks 통합
