--- src/features/comment/api/commentApi.ts
@@ -1,36 +1,0 @@
-import { CommentType } from "@/entities/comment/model/types";
-import {
-  PostCommentRequestType,
-  PostCommentResponseType,
-} from "../model/types";
-import { PutCommentRequestType } from "../model/types";
-import { createApi } from "@/shared/api/baseApi";
-
-const api = createApi("/api/comments");
-
-export const commentApi = {
-  // 댓글 추가
-  async createComment(
-    commentData: PostCommentRequestType,
-  ): Promise<PostCommentResponseType> {
-    return api.post<PostCommentResponseType>("/add", commentData);
-  },
-
-  // 댓글 수정
-  async updateComment(
-    id: number,
-    commentData: Pick<PutCommentRequestType, "body">,
-  ): Promise<CommentType> {
-    return api.put<CommentType>(`/${id}`, commentData);
-  },
-
-  // 댓글 삭제
-  async deleteComment(id: number): Promise<void> {
-    return api.delete(`/${id}`);
-  },
-
-  // 댓글 좋아요
-  async likeComment(id: number, currentLikes: number): Promise<CommentType> {
-    return api.patch<CommentType>(`/${id}`, { likes: currentLikes + 1 });
-  },
-};
