+++ src/features/comment/api/index.ts
@@ -1,2 +1,1 @@
 export * from "./hooks";
-export * from "./commentApi";
